Languages duplicated
When the server starts we have run through to cache the routes
and that is being added to the array we declared and when we visit
the route the array is being add once again so we have duplicate data.
This will allow us to be able to reuse what we already have.